+++ controllers/nifiparametercontext_controller.go
@@ -128,6 +128,27 @@ func (r *NifiParameterContextReconciler) Reconcile(ctx context.Context, req ctrl
 		parameterSecrets = append(parameterSecrets, secret)
 	}
 
+	// Get the referenced NiFiParameterContext referenced
+	var parameterContextRefs []*v1alpha1.NifiParameterContext
+	for _, parameterContextRef := range instance.Spec.InheritedParameterContexts {
+		parameterContextNamespace := GetParameterContextRefNamespace(instance.Namespace, parameterContextRef)
+		var parameterContext *v1alpha1.NifiParameterContext
+		if parameterContext, err = k8sutil.LookupNifiParameterContext(r.Client, parameterContextRef.Name, parameterContextNamespace); err != nil {
+			// This shouldn't trigger anymore, but leaving it here as a safetybelt
+			if k8sutil.IsMarkedForDeletion(instance.ObjectMeta) {
+				r.Log.Info("Secret is already gone, there is nothing we can do")
+				if err = r.removeFinalizer(ctx, instance); err != nil {
+					return RequeueWithError(r.Log, "failed to remove finalizer", err)
+				}
+				return Reconciled()
+			}
+
+			// the cluster does not exist - should have been caught pre-flight
+			return RequeueWithError(r.Log, "failed to lookup referenced parameter context", err)
+		}
+		parameterContextRefs = append(parameterContextRefs, parameterContext)
+	}
+
 	// Prepare cluster connection configurations
 	var clientConfig *clientconfig.NifiConfig
 	var clusterConnect clientconfig.ClusterConnect
@@ -187,7 +208,7 @@ func (r *NifiParameterContextReconciler) Reconcile(ctx context.Context, req ctrl
 
 	// Check if marked for deletion and if so run finalizers
 	if k8sutil.IsMarkedForDeletion(instance.ObjectMeta) {
-		return r.checkFinalizers(ctx, instance, parameterSecrets, clientConfig)
+		return r.checkFinalizers(ctx, instance, parameterSecrets, parameterContextRefs, clientConfig)
 	}
 
 	// Ensure the cluster is ready to receive actions
@@ -206,7 +227,7 @@ func (r *NifiParameterContextReconciler) Reconcile(ctx context.Context, req ctrl
 	// Ìn case of the cluster reference changed.
 	if !v1alpha1.ClusterRefsEquals([]v1alpha1.ClusterReference{instance.Spec.ClusterRef, current.Spec.ClusterRef}) {
 		// Delete the resource on the previous cluster.
-		if err := parametercontext.RemoveParameterContext(instance, parameterSecrets, clientConfig); err != nil {
+		if err := parametercontext.RemoveParameterContext(instance, parameterSecrets, parameterContextRefs, clientConfig); err != nil {
 			r.Recorder.Event(instance, corev1.EventTypeWarning, "RemoveError",
 				fmt.Sprintf("Failed to delete NifiParameterContext %s from cluster %s before moving in %s",
 					instance.Name, original.Spec.ClusterRef.Name, original.Spec.ClusterRef.Name))
@@ -249,7 +270,7 @@ func (r *NifiParameterContextReconciler) Reconcile(ctx context.Context, req ctrl
 		}
 		if status == nil {
 			// Create NiFi parameter context
-			status, err = parametercontext.CreateParameterContext(instance, parameterSecrets, clientConfig)
+			status, err = parametercontext.CreateParameterContext(instance, parameterSecrets, parameterContextRefs, clientConfig)
 			if err != nil {
 				return RequeueWithError(r.Log, "failure creating parameter context "+instance.Name, err)
 			}
@@ -267,7 +288,7 @@ func (r *NifiParameterContextReconciler) Reconcile(ctx context.Context, req ctrl
 	// Sync ParameterContext resource with NiFi side component
 	r.Recorder.Event(instance, corev1.EventTypeNormal, "Synchronizing",
 		fmt.Sprintf("Synchronizing parameter context %s", instance.Name))
-	status, err := parametercontext.SyncParameterContext(instance, parameterSecrets, clientConfig)
+	status, err := parametercontext.SyncParameterContext(instance, parameterSecrets, parameterContextRefs, clientConfig)
 	if status != nil {
 		instance.Status = *status
 		if err := r.Client.Status().Update(ctx, instance); err != nil {
@@ -353,12 +374,13 @@ func (r *NifiParameterContextReconciler) checkFinalizers(
 	ctx context.Context,
 	parameterContext *v1alpha1.NifiParameterContext,
 	parameterSecrets []*corev1.Secret,
+	parameterContextRefs []*v1alpha1.NifiParameterContext,
 	config *clientconfig.NifiConfig) (reconcile.Result, error) {
 	r.Log.Info("NiFi parameter context is marked for deletion. Removing finalizers.",
 		zap.String("parameterContext", parameterContext.Name))
 	var err error
 	if util.StringSliceContains(parameterContext.GetFinalizers(), parameterContextFinalizer) {
-		if err = r.finalizeNifiParameterContext(parameterContext, parameterSecrets, config); err != nil {
+		if err = r.finalizeNifiParameterContext(parameterContext, parameterSecrets, parameterContextRefs, config); err != nil {
 			return RequeueWithError(r.Log, "failed to finalize parameter context "+parameterContext.Name, err)
 		}
 		if err = r.removeFinalizer(ctx, parameterContext); err != nil {
@@ -379,9 +401,10 @@ func (r *NifiParameterContextReconciler) removeFinalizer(ctx context.Context, pa
 func (r *NifiParameterContextReconciler) finalizeNifiParameterContext(
 	parameterContext *v1alpha1.NifiParameterContext,
 	parameterSecrets []*corev1.Secret,
+	parameterContextRefs []*v1alpha1.NifiParameterContext,
 	config *clientconfig.NifiConfig) error {
 
-	if err := parametercontext.RemoveParameterContext(parameterContext, parameterSecrets, config); err != nil {
+	if err := parametercontext.RemoveParameterContext(parameterContext, parameterSecrets, parameterContextRefs, config); err != nil {
 		return err
 	}
 	r.Log.Info("Deleted NifiParameter Context",